keep spark api binary compatible
restore int signatures and add compatibility overload

+++ spark-integration/src/main/scala/io/toonformat/toon4s/spark/monitoring/ToonMonitoring.scala
@@ -111,11 +111,11 @@ object ToonMonitoring {
 
   final case class EncodingMetrics(
       encodingTimeMs: Long,
-      jsonTokenCount: Long,
-      toonTokenCount: Long,
+      jsonTokenCount: Int,
+      toonTokenCount: Int,
       savingsPercent: Double,
       chunkCount: Int,
-      avgChunkSize: Long,
+      avgChunkSize: Int,
       success: Boolean,
       errorType: Option[String],
   )
